reworked clone system call to work with system call emulation.
