+++ .vscode/launch.json
@@ -348,6 +348,8 @@
          "type": "pickString",
          "options": [
             "sigaltstack_test",
+            "clone_test",
+            "socket_test",
             "pthread_cancel_test",
             "var_storage_test",
             "exit_value_test",
