+++ km/gdb/list.gdb
@@ -39,8 +39,8 @@ if $argc == 0
    while $item != 0
       set $distance = $item->start - $last_end
       printf "%3d 0x%-12lx next 0x%-12lx start= 0x%lx size = 0x%05x (%08ld) prot 0x%x flags 0x%x fn %s km_fl 0x%x distance %d (%ld)\n", \
-               $count, $item, $item->link->tqe_next, $item->start/4096ul, $item->size/4096ul, \
-               $item->size, $item->protection, $item->flags, $item->filename, $item->km_flags.data32, $distance/4096ul, $distance
+               $count, $item, $item->link->tqe_next, $item->start/0x1000ul, $item->size/0x1000ul, $item->size, \
+               $item->protection, $item->flags, $item->filename, $item->km_flags.data32, $distance/0x1000ul, $distance
       set $count++
       set $last_end = $item->start + $item->size
       set $item=$item->link->tqe_next
