+++ km/km_vcpu_run.c
@@ -404,7 +404,7 @@ static int hypercall(km_vcpu_t* vcpu, int* hc)
               r->io.size,
               r->io.direction == KVM_EXIT_IO_OUT ? "out" : "in");
 
-      siginfo_t info = {.si_signo = SIGSYS, .si_code = SI_KERNEL};
+      siginfo_t info = {.si_signo = SIGBUS, .si_code = SI_KERNEL};
       km_post_signal(vcpu, &info);
       return -1;
    }
@@ -446,7 +446,7 @@ static int hypercall(km_vcpu_t* vcpu, int* hc)
    km_kma_t ga_kma;
    if ((ga_kma = km_gva_to_kma(ga)) == NULL || km_gva_to_kma(ga + sizeof(km_hc_args_t) - 1) == NULL) {
       km_infox(KM_TRACE_SIGNALS, "hc: %d bad km_hc_args_t address:0x%lx", *hc, ga);
-      siginfo_t info = {.si_signo = SIGSYS, .si_code = SI_KERNEL};
+      siginfo_t info = {.si_signo = SIGSEGV, .si_code = SI_KERNEL};
       km_post_signal(vcpu, &info);
       return -1;
    }
