+++ tests/clone_test.c
@@ -15,7 +15,7 @@
 
 static int childFunc(void* arg)
 {
-   printf("Hello from clone\n");
+   fprintf(stderr, "Hello from clone\n");
    return 0; /* Child terminates now */
 }
 
@@ -25,21 +25,19 @@ int main(int argc, char* argv[])
 {
    unsigned flags = CLONE_VM | CLONE_FS | CLONE_FILES | CLONE_SIGHAND | CLONE_THREAD |
                     CLONE_SYSVSEM | CLONE_PARENT_SETTID | CLONE_CHILD_CLEARTID | CLONE_DETACHED;
-   char* stack;    /* Start of stack buffer */
-   char* stackTop; /* End of stack buffer */
+   char* stack;     /* Start of stack buffer */
    pid_t pid, tid;
 
    stack = malloc(STACK_SIZE);
-   if (stack == NULL)
+   if (stack == NULL) {
       errExit("malloc");
-   stackTop = stack + STACK_SIZE - 16; /* Assume stack grows downward */
+   }
 
    printf("clone()\n");
-
-   pid = clone(childFunc, stackTop, flags, NULL, &pid, NULL, &tid);
-   if (pid == -1)
+   pid = clone(childFunc, stack + STACK_SIZE, flags, NULL, &pid, NULL, &tid);
+   if (pid == -1) {
       errExit("clone");
-
+   }
    printf("clone() returned %ld\n", (long)pid);
 
    usleep(1000);
